ValidationAttemptJQuery_CW_3_30
Having issues with the getElementById constantly returning 'null'. I have enclosed the script with 'document.ready' which is suppose to only act when the DOM is ready; however, it's still null even after textbox entry. Still researching on why this is happening. :( (I've attempted to this with the 'id' of Customer_Name that is created within the Razor and an overloaded 'id' property within the HTML attributes. Both are still showing Null.)